requirements.txt
torch>=1.9.0
numpy>=1.19.0
matplotlib>=3.3.0
scipy>=1.7.0
pyyaml>=5.4.0
pytest>=6.0.0
jupyter>=1.0.0
tensorboard>=2.7.0
tqdm>=4.62.0